refactor: favorites service 패키지 분리
